extract KeyInfoInteractor from KeyLoader
